pytrade/requirements.txt
requests~=2.31.0
pandas~=2.1.2
python-dotenv~=1.0.0
mysql-connector-python~=8.2.0

